requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.6.1
Automat==0.7.0
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
numpy==1.16.4
Pillow==6.0.0
pycparser==2.19
PyHamcrest==1.9.0
six==1.12.0
Twisted==19.2.0
txaio==18.8.1
wxPython==4.0.6
zope.interface==4.6.0
pydocker
